% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSwap.R
\name{oneswap}
\alias{oneswap}
\title{Title}
\usage{
oneswap(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
True if vector can be ordered with at most one swap
}
\description{
Title
}
\examples{
 oneswap(c(1,2,3,4,5))



}
